Hack Reactor halfway hackthon
Let's celebrate with a hackathon! A hackathon is a great opportunity to work on a larger, real-world project, while reviewing and solidifying the core concepts that we've covered in the first half of the courseStep 1: Make a Plan

Take a few minutes to brainstorm some project ideas. This could be related to your community site, a stand-alone project, or just a crazy mash-up site to help you practice different coding concepts
Review the Step 3: Coding section below to consider the different concepts that we hope to practice during this project.
In groups of four, hold a standup meeting to present your plan of what you'd like to accomplish, and potential stumbling blocks.
If someone mentions being concerned about a concept that you feel solid on, offer to be a resource for them if they have questions